Move existing geo_coordinate methods to geo provider
